renamed data package to reader
