+++ sciencebeam_trainer_delft/grobid_trainer.py
@@ -18,7 +18,7 @@ from sciencebeam_trainer_delft.wrapper import Sequence
 from sciencebeam_trainer_delft.cloud_support import patch_cloud_support
 from sciencebeam_trainer_delft.embedding_manager import EmbeddingManager
 from sciencebeam_trainer_delft.models import get_model_names, patch_get_model
-from sciencebeam_trainer_delft.data import load_data_and_labels_crf_file
+from sciencebeam_trainer_delft.reader import load_data_and_labels_crf_file
 from sciencebeam_trainer_delft.download_manager import DownloadManager
 from sciencebeam_trainer_delft.utils import get_tf_info
